# Reuse allocated buffer space in merge iterator.