@bboreham 2023-05-26
I think it would be good to run the existing benchmarks too, just to check that `len-1` isn't unexpectedly expensive.