@chaudum 2023-05-26
Quickly ran existing benchmarks:

```
$ gotest -v -count=10 ./pkg/iter/... -test.run="^$" -test.bench=Benchmark -test.benchtime=2000x
```

```console
$ benchstat mergeiter-old.txt mergeiter-new.txt                                                                                      
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/iter
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                           │ mergeiter-old.txt │         mergeiter-new.txt          │
                           │      sec/op       │   sec/op     vs base               │
SortIterator/merge_sort-8          742.8µ ± 1%   718.8µ ± 1%  -3.24% (p=0.000 n=10)
SortIterator/sort-8                632.9µ ± 4%   608.6µ ± 2%  -3.84% (p=0.001 n=10)
SortSampleIterator/merge-8         1.554m ± 2%   1.529m ± 3%  -1.63% (p=0.035 n=10)
SortSampleIterator/sort-8          274.5µ ± 8%   274.9µ ± 7%       ~ (p=0.853 n=10)
geomean                            669.2µ        654.8µ       -2.15%

                          │ mergeiter-old.txt │           mergeiter-new.txt           │
                          │       B/op        │     B/op      vs base                 │
SortIterator/merge_sort-8        21.22Ki ± 0%   21.22Ki ± 0%       ~ (p=1.000 n=10) ¹
SortIterator/sort-8              13.94Ki ± 0%   13.94Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                          17.20Ki        17.20Ki       +0.00%
¹ all samples are equal

                          │ mergeiter-old.txt │          mergeiter-new.txt          │
                          │     allocs/op     │ allocs/op   vs base                 │
SortIterator/merge_sort-8          7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=10) ¹
SortIterator/sort-8                5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                            5.916        5.916       +0.00%

```